Move sidebar container to its own folder.
