+++ src/sidebar-common/comment-box/components/tag-input-container.tsx
@@ -3,7 +3,8 @@ import { connect, MapStateToProps } from 'react-redux'
 
 import * as actions from '../actions'
 import * as selectors from '../selectors'
-import State, { MapDispatchToProps } from '../../types'
+import State from '../types'
+import { MapDispatchToProps } from '../../types'
 import { TagInput } from '../../components'
 
 interface StateProps {
