+++ src/sidebar-common/components/tag-input.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import { IndexDropdown } from '../../common-ui/containers'
+import { IndexDropdown } from 'src/common-ui/containers'
 import TagHolder from './tag-holder'
 
 interface Props {
