+++ src/sidebar-common/sidebar/index.ts
@@ -1,0 +1,8 @@
+import SidebarContainer from './sidebar-container'
+import * as actions from './actions'
+import reducer from './reducer'
+import * as selectors from './selectors'
+import State, * as types from './types'
+
+export default SidebarContainer
+export { actions, reducer, selectors, State, types }
