+++ src/sidebar-common/sidebar/reducer.ts
@@ -6,8 +6,8 @@ import State, { Annotation, Page } from './types'
 import {
     reducer as commentBoxReducer,
     defaultState as defCommentBoxState,
-} from './comment-box'
-import AnnotationsManager from './annotations-manager'
+} from '../comment-box'
+import AnnotationsManager from '../annotations-manager'
 
 export const defaultState: State = {
     annotationsManager: null,
