+++ src/sidebar-overlay/index.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import * as ReactDOM from 'react-dom'
 
 import RibbonSidebarController from './ribbon-sidebar-controller'
-import AnnotationsManager from '../sidebar-common/annotations-manager'
+import AnnotationsManager from 'src/sidebar-common/annotations-manager'
 import {
     highlightAnnotations,
     highlightAndScroll,
