+++ src/sidebar-overlay/reducer.ts
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux'
 
 import RootState from './types'
 import { reducer as ribbon } from './ribbon'
-import { reducer as sidebar } from '../sidebar-common'
+import { reducer as sidebar } from 'src/sidebar-common'
 
 const rootReducer = combineReducers<RootState>({
     ribbon,
